# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DirectoryImportApplyResult": {
        "properties": {
          "applied": {
            "type": "boolean"
          },
          "preview": {
            "$ref": "#/components/schemas/DirectoryImportPreviewResult"
          },
          "syncRunId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "applied",
          "preview",
          "syncRunId"
        ],
        "type": "object"
      },
      "DirectoryImportChange": {
        "properties": {
          "action": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "selectable": {
            "type": "boolean"
          },
          "simulationCode": {
            "type": "string"
          },
          "sourceRecordId": {
            "type": "string"
          }
        },
        "required": [
          "action",
          "displayName",
          "selectable",
          "simulationCode",
          "sourceRecordId"
        ],
        "type": "object"
      },
      "DirectoryImportIssue": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "rowNumber": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "null",
              "string"
            ]
          },
          "sourceRecordId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "rowNumber",
          "sourceRecordId"
        ],
        "type": "object"
      },
      "DirectoryImportPreviewResult": {
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/DirectoryImportChange"
            },
            "type": "array"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/DirectoryImportIssue"
            },
            "type": "array"
          },
          "insertCount": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "parsedPractitionerCount": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "previewToken": {
            "default": "",
            "type": "string"
          },
          "rejectedCount": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "sourceSystem": {
            "type": "string"
          },
          "updateCount": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/DirectoryImportIssue"
            },
            "type": "array"
          }
        },
        "required": [
          "changes",
          "errors",
          "insertCount",
          "parsedPractitionerCount",
          "rejectedCount",
          "sourceSystem",
          "updateCount",
          "warnings"
        ],
        "type": "object"
      },
      "DirectoryPractitionerListItem": {
        "properties": {
          "availableChannels": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "department": {
            "type": [
              "null",
              "string"
            ]
          },
          "displayName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "isStale": {
            "type": "boolean"
          },
          "lastName": {
            "type": "string"
          },
          "lastSynchronizedAtUtc": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "onCallLastSynchronizedAtUtc": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "onCallSourceSystem": {
            "type": [
              "null",
              "string"
            ]
          },
          "onCallTier": {
            "type": [
              "null",
              "string"
            ]
          },
          "practitionerId": {
            "format": "uuid",
            "type": "string"
          },
          "practitionerRoleId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "roleTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "selectable": {
            "type": "boolean"
          },
          "selectionRevision": {
            "type": "string"
          },
          "simulationCode": {
            "type": "string"
          },
          "site": {
            "type": [
              "null",
              "string"
            ]
          },
          "sourceSystem": {
            "type": [
              "null",
              "string"
            ]
          },
          "specialty": {
            "type": "string"
          }
        },
        "required": [
          "availableChannels",
          "department",
          "displayName",
          "firstName",
          "isActive",
          "isStale",
          "lastName",
          "lastSynchronizedAtUtc",
          "onCallLastSynchronizedAtUtc",
          "onCallSourceSystem",
          "onCallTier",
          "practitionerId",
          "practitionerRoleId",
          "roleTitle",
          "selectable",
          "selectionRevision",
          "simulationCode",
          "site",
          "sourceSystem",
          "specialty"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Simulation-only contract. It is not a hospital or clinical API.",
    "title": "Critical Alerts Simulation API",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/directory/imports": {
      "post": {
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  },
                  "preview_token": {
                    "type": "string"
                  }
                },
                "required": [
                  "file",
                  "preview_token"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryImportApplyResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryImportApplyResult"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "413": {
            "description": "Payload Too Large"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "DirectoryEndpoints"
        ]
      }
    },
    "/api/v1/directory/imports/preview": {
      "post": {
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryImportPreviewResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "413": {
            "description": "Payload Too Large"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "DirectoryEndpoints"
        ]
      }
    },
    "/api/v1/directory/practitioners": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "department",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "site",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "onCallNow",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "includeInactive",
            "schema": {
              "default": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DirectoryPractitionerListItem"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "DirectoryEndpoints"
        ]
      }
    },
...
  },
...
}